backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
phi-agent==0.1.0
python-dotenv==1.0.0